+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Added `parse_connection_string_to_dict` function for generic parsing and validation of connection strings across SDKs.
+- Added `azure.core.utils.parse_connection_string` function to parse connection strings across SDKs, with common validation and support for case insensitive keys.
 
 ## 1.12.0 (2021-03-08)